test: split media upgrade coverage
